pyproject.toml
[project]
name = "devin-acu-crawler"
version = "0.1.0"
description = "Get Devin ACUs"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pandas>=3.0.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
